refactor: button component
